src/service/requirements.txt
flask
pymysql
gunicorn
dotenv
click
flask_jwt
flask_sqlalchemy
flask_migrate
flask_marshmallow
marshmallow-sqlalchemy
mysql
mysqlclient